+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `to_json` methods for `config::ConfigSpaceAnalysisResults` and `config::DoFSpaceAnalysisResults`.
 - Added `Prim.__repr__`, `Supercell.__repr__`, `SupercellRecord.__repr__`, `Configuration.__repr__`, `ConfigurationRecord.__repr__`, `ConfigurationWithProperties.__repr__`, `DoFSpaceAnalysisResults.__repr__`.
 - Added `ConfigSpaceAnalysisResults.to_dict` and `DoFSpaceAnalysisResults.to_dict` methods.
+- Added `OccSystem.__repr__`, `OccPosition.copy`, `OccPosition.__copy__`, `OccPosition.__deepcopy__`, `OccPosition.__repr__`, `OccEventRep.__repr__`, `OccEventRep.copy`, `OccEventRep.__copy__`, `OccEventRep.__deepcopy__`, `OccEvent.__repr__`.
 
 
 ## [2.0a4] - 2024-07-16